[FIX] Temp Update on Material Assignment (#420)
* Temp Update on Material Assignment

Previously it was shader assignment by request or iterative order from URP to Unlit, but LLM usually assign Standard in a URP setting.

This small fix introduce the helper to resolve some of the conflicts, while give warnings in the log to show if the user is creating a shader in a non-compatible setting.

* Update RenderPipelineUtility.cs